requirements.txt
arrow==1.2.3
boto3==1.26.56
botocore==1.29.56
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
idna==3.4
jmespath==1.0.1
marshmallow==3.19.0
packaging==23.0
pyTenable==1.4.11
python-box==6.1.0
python-dateutil==2.8.2
requests==2.28.2
restfly==1.4.7
s3transfer==0.6.0
semver==2.13.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.14
